geanynumberedbookmarks: Don't let used key presses slip through
For some reason the plugin uses key release events rather than key
presses, but most other consumers (including Geany itself) uses
presses.  This means that if we don't report the press event as
consumed, hitting the combo might result in an action happening on the
key press, and ours on the key release, which is fairly surprising.

To avoid this, consume the relevant key presses as well, doing nothing.